api/requirements.txt
tensorflow==2.7.0
python-multipart
fastapi
uvicorn
pillow
tensorflow
numpy
matplotlib